+++ OrangeShopBE/src/main/java/com/orange/payload/request/UpdateOrderStatus.java
@@ -5,5 +5,5 @@ import lombok.Data;
 @Data
 public class UpdateOrderStatus {
     private Long orderId;
-    private Integer orderStatusId;
+    private Long orderStatusId;
 }
